BACKEND/attendance/requirements.txt
fastapi
strawberry-graphql
uvicorn
sqlalchemy
asyncpg
psycopg2
pydantic
httpx
python-dotenv
